icap-v7-master/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg==3.2.3
gunicorn==23.0.0
requests==2.32.3
redis==5.2.1
channels[daphne]==4.2.0
channels-redis==4.2.1
django-rest-knox==5.0.2
django-import-export==4.3.3
django-extensions==3.2.3
drf-spectacular[sidecar]==0.28.0
jupyter==1.1.1
ipython==8.31.0
notebook==7.3.2
django-colorfield==0.11.0
celery==5.4.0
pandas==2.2.3
XlsxWriter==3.2.3
openpyxl==3.1.5
pika==1.3.2
Flask==3.1.0
beautifulsoup4==4.12.2
pdfkit==1.0.0
extract_msg==0.52.0
django-auth-ldap==5.1.0
PyPDF2==3.0.1
numpy==2.2.5
pdf2image==1.17.0
pdfplumber==0.11.5
opencv-python==4.11.0.86
python-doctr==0.11.0
easyocr==1.7.2
PyMuPDF==1.25.3
websockets==15.0.1
asgiref==3.8.1
python-dotenv==1.1.0
xlrd==2.0.1
lxml==5.4.0
PyYAML==6.0.2
pycryptodome==3.20
qdrant-client==1.15.1
sentence-transformers==5.1.0
cryptography==45.0.7
pgvector==0.4.1
Pillow==11.1.0